full norminette done
