+++ src/calling_the_execs_shell.c
@@ -88,7 +88,7 @@ void	parse_and_exec_cmd_shell(char **cmd, char **env)
 	}
 	cmd[0] = ft_strjoin_arena("/", cmd[0]);
 	i = 0;
-	while (p.path[i])
+	while (p.path[i++])
 	{
 		if (search_path_exec(p.path[i], cmd[0]) == true)
 		{
@@ -96,7 +96,6 @@ void	parse_and_exec_cmd_shell(char **cmd, char **env)
 			execve(cmd[0], cmd, env);
 			exit(1);
 		}
-		i++;
 	}
 	fprintf(stderr, "%s: command not found\n", cmd[0] + 1);
 	free_strrarr(p.path);
