+++ src/parsing_2.c
@@ -79,13 +79,9 @@ char	*del_quotes(t_parsing *parse_list, char *line)
 	newline = arena_alloc(&g_ex.arena, ft_strlen(line) * sizeof(char *));
 	while (line[i] != '\0')
 	{
-		if ((parse_list->quote_type == 39 && line[i] == parse_list->quote_type)
-			&& (i != parse_list->quote_start && i != parse_list->quote_end))
-		{
-			newline[y] = line[i];
-			y++;
-		}
-		else if (line[i] != parse_list->quote_type)
+		if (((parse_list->quotes == true && line[i] == parse_list->quote_type)
+				&& (i != parse_list->quote_start && i != parse_list->quote_end))
+			|| (line[i] != parse_list->quote_type))
 		{
 			newline[y] = line[i];
 			y++;
